requirements.txt
# PyTorch và related packages - phiên bản tương thích
torch>=1.12.0,<2.0.0
torchvision>=0.13.0,<1.0.0
torchaudio>=0.12.0,<1.0.0

# Medical imaging libraries
nibabel>=3.2.1
SimpleITK>=2.1.1
pydicom>=2.3.0

# Image processing
opencv-python>=4.5.0
scikit-image>=0.19.0
Pillow>=8.3.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Machine learning và visualization
matplotlib>=3.5.0
tensorboard>=2.8.0
tqdm>=4.62.0

# Development tools
pytest>=6.2.0
black>=21.9.0 